Refactored Nav elements to use separate components
